Added caching till the end of the day if an update from today is detected
After running a check till next update for about 10 days I found that the ge
really updates on seemingly random times throughout the day. But never twice
on the same day. So if we detect that the last reported date in the graph is
today we simply cache till the end of today (UTC). Otherwise we'll cache 1 minute
to at least prevent multiple calls to the actual api.